--- a/app/services/school_data_importer.rb
+++ b/app/services/school_data_importer.rb
@@ -14,27 +14,17 @@ class SchoolDataImporter
 
   def run
     files = school_data_files
+    la_cache = {}
+    lad_cache = {}
 
-    CSV.foreach(files["ecf_tech.csv"].path, headers: true, encoding: "ISO-8859-1:UTF-8") do |row|
-      school = row_to_school(row)
-      local_authority = row_to_local_authority(row)
-      local_authority_district = row_to_lad(row)
-
-      link_school_to_local_authority(school, local_authority)
-      link_school_to_lad(school, local_authority_district)
-    end
-  end
-
-  def update_emails
-    files = school_data_files
-
-    CSV.foreach(files["ecf_tech.csv"].path, headers: true, encoding: "ISO-8859-1:UTF-8") do |row|
-      school = School.find_by(urn: row.fetch("URN"))
-      next if school.blank?
-
-      school.primary_contact_email = row.fetch("MainEmail")
-      school.secondary_contact_email = row.fetch("AlternativeEmail")
-      school.save!
+    CSV.foreach(files["ecf_tech.csv"].path, headers: true, encoding: "ISO-8859-1:UTF-8").with_index do |row|
+      SchoolHandler.new(
+        data: row,
+        logger: logger,
+        start_year: start_year,
+        la_cache: la_cache,
+        lad_cache: lad_cache
+      ).call
     end
   end
 
@@ -46,77 +36,123 @@ private
     @gias_files
   end
 
-  def row_to_school(row)
-    school = School.find_or_initialize_by(urn: row.fetch("URN"))
-    school.name = row.fetch("EstablishmentName")
-    school.school_type_code = row.fetch("TypeOfEstablishment (code)").to_i
-    school.school_type_name = row.fetch("TypeOfEstablishment (name)")
-    school.address_line1 = row.fetch("Street")
-    school.address_line2 = row.fetch("Locality")
-    school.address_line3 = row.fetch("Town")
-    school.postcode = row.fetch("Postcode")
-    school.ukprn = row.fetch("UKPRN")
-    school.school_phase_type = row.fetch("PhaseOfEducation (code)").to_i
-    school.school_phase_name = row.fetch("PhaseOfEducation (name)")
-    school.school_website = row.fetch("SchoolWebsite")
-    school.school_status_code = row.fetch("EstablishmentStatus (code)").to_i
-    school.school_status_name = row.fetch("EstablishmentStatus (name)")
-    school.administrative_district_code = row.fetch("DistrictAdministrative (code)")
-    school.administrative_district_name = row.fetch("DistrictAdministrative (name)")
-    school.primary_contact_email = row.fetch("MainEmail")
-    school.secondary_contact_email = row.fetch("AlternativeEmail")
-    school.save!
-    school
-  end
+  class SchoolHandler
+    attr_reader :school, :data, :logger, :start_year
 
-  def row_to_local_authority(row)
-    local_authority = LocalAuthority.find_or_initialize_by(code: row.fetch("LA (code)"))
-    row_local_authority_name = row.fetch("LA (name)")
+    def initialize(data:, logger:, start_year:, la_cache:, lad_cache:)
+      @logger = logger
+      @data = data
+      @start_year = start_year
+      @la_cache = la_cache
+      @lad_cache = lad_cache
 
-    if local_authority.persisted? && local_authority.name != row_local_authority_name
-      logger.info "LA name change in school import. Old name: #{local_authority.name}, New name: #{row_local_authority_name}"
+      find_school
     end
 
-    local_authority.name = row_local_authority_name
-    local_authority.save!
-    local_authority
-  end
+    def call
+      ActiveRecord::Base.transaction do
+        record_school_details
+        return unless new_school?
 
-  def row_to_lad(row)
-    local_authority_district = LocalAuthorityDistrict.find_or_initialize_by(code: row.fetch("DistrictAdministrative (code)"))
-    row_lad_name = row.fetch("DistrictAdministrative (name)")
-
-    if local_authority_district.persisted? && local_authority_district.name != row_lad_name
-      logger.info "LA name change in school import. Old name: #{local_authority_district.name}, New name: #{row_lad_name}"
+        link_school_to_local_authority
+        link_school_to_local_district_authority
+      end
     end
 
-    local_authority_district.name = row_lad_name
-    local_authority_district.save!
-    local_authority_district
-  end
+  private
 
-  def link_school_to_local_authority(school, local_authority)
-    school_local_authority = SchoolLocalAuthority.find_or_initialize_by(school: school, local_authority: local_authority)
-
-    if school_local_authority.new_record?
-      school_local_authority.start_year = start_year
-      SchoolLocalAuthority.latest.find_by(school: school)&.update!(end_year: start_year)
+    def new_school?
+      @new_school
     end
 
-    school_local_authority.save!
-  end
-
-  def link_school_to_lad(school, lad)
-    school_lad = SchoolLocalAuthorityDistrict.find_or_initialize_by(
-      school: school,
-      local_authority_district: lad,
-    )
-
-    if school_lad.new_record?
-      school_lad.start_year = start_year
-      SchoolLocalAuthorityDistrict.latest.find_by(school: school)&.update!(end_year: start_year)
+    def find_school
+      @school = School.find_or_initialize_by(urn: data.fetch("URN"))
+      @new_school = @school.new_record?
     end
 
-    school_lad.save!
+    def record_school_details
+      school.name = data.fetch("EstablishmentName")
+      school.address_line1 = data.fetch("Street")
+      school.address_line2 = data.fetch("Locality")
+      school.address_line3 = data.fetch("Town")
+      school.postcode = data.fetch("Postcode")
+      school.ukprn = data.fetch("UKPRN")
+      school.school_phase_type = data.fetch("PhaseOfEducation (code)").to_i
+      school.school_phase_name = data.fetch("PhaseOfEducation (name)")
+      school.school_website = data.fetch("SchoolWebsite")
+      school.primary_contact_email = data.fetch("MainEmail").presence
+      school.secondary_contact_email = data.fetch("AlternativeEmail").presence
+
+      # Changes to properties below carries major consequences and must be avoided until we know how to handle them
+      if new_school?
+        school.school_type_code = data.fetch("TypeOfEstablishment (code)").to_i
+        school.school_type_name = data.fetch("TypeOfEstablishment (name)")
+
+        school.school_status_code = data.fetch("EstablishmentStatus (code)").to_i
+        school.school_status_name = data.fetch("EstablishmentStatus (name)")
+
+        school.administrative_district_code = data.fetch("DistrictAdministrative (code)")
+        school.administrative_district_name = data.fetch("DistrictAdministrative (name)")
+      end
+
+      school.save!
+    end
+
+    def local_authority
+      return @local_authority if @local_authority
+
+      code = data.fetch("LA (code)")
+      @la_cache.fetch(code) do
+        local_authority = LocalAuthority.find_or_initialize_by(code: code)
+        new_name = data.fetch("LA (name)")
+
+        if local_authority.name != new_name
+          if local_authority.persisted?
+            logger.info "LA name changed in school import. Old name: #{local_authority.name}, New name: #{new_name}"
+          end
+          local_authority.name = new_name
+          local_authority.save!
+        end
+
+        @local_authority = @la_cache[code] = local_authority
+      end
+    end
+
+    def local_district_authority
+      return @local_district_authority if @local_district_authority
+
+      code = data.fetch("DistrictAdministrative (code)")
+
+      @lad_cache.fetch(code) do
+        local_authority_district = LocalAuthorityDistrict.find_or_initialize_by(code: code)
+        new_name = data.fetch("DistrictAdministrative (name)")
+
+        if local_authority_district.name != new_name
+          if local_authority_district.persisted?
+            logger.info "LAD name changed in school import. Old name: #{local_authority_district.name}, New name: #{new_name}"
+          end
+          local_authority_district.name = new_name
+          local_authority_district.save!
+        end
+
+        @local_district_authority = @lad_cache[code] = local_authority_district
+      end
+    end
+
+    def link_school_to_local_authority
+      SchoolLocalAuthority.create!(
+        school: school,
+        local_authority: local_authority,
+        start_year: start_year,
+      )
+    end
+
+    def link_school_to_local_district_authority
+      SchoolLocalAuthorityDistrict.create!(
+        school: school,
+        local_authority_district: local_district_authority,
+        start_year: start_year,
+      )
+    end
   end
 end