--- a/lib/tasks/cron.rake
+++ b/lib/tasks/cron.rake
@@ -4,6 +4,6 @@ namespace :cron do
   desc "Updates DelayedJob cron-scheduled tasks"
   task schedule: :environment do
     SessionTrimJob.schedule
-    SchoolEmailUpdateJob.schedule
+    SchoolDataImporterJob.schedule
   end
 end