… app/jobs/school_email_update_job.rb
… spec/jobs/school_email_update_job_spec.rb