https://dfedigital.atlassian.net/browse/CPDRP-388

This PR modifies SchoolDataImport to ignore certain changes in school information for existing schools. This is a temporary change until we can handle those problematic changes.

Additionally I have applied few performance tweaks, reducing the  execution time on the empty database from 1118s to 465s and each other run from 562s to 36s (however removing LA and LAD association handling had quite a bit impact here as well). These are:
1) Loading schools in batches of 1000
2) Caching LocalAuthority and LocalAuthorityDistrict records

Some further speed up could bo achieved by:
1) Creating a thread pool and handling schools concurrenty (requires thread safe cache object)
2) Performing batch inserts