@ethanmills 2021-06-18
Eugh, totally misunderstood the code. Ignore my request for changes, still reading